Check that the m_closedShape has a size > 0 before using the iterator
An iterator on m_closedShape is being created and used without checking
if the iterator actually has a valid value. This was not a problem in
Qt5, but in Qt6 causes a crash. This patch adds a check to the QMap, so
that the iterator is not even created if the QMap has a size of 0

Change-Id: Ie7a3ce2ad162bb787fbbb18674bbf9bd95524453
Reviewed-by: Rebecca Worledge <rebecca.worledge@qt.io>