Update commercial license headers
Updated header.COMM to the files in tqtc-qtlottie. Examples, tests
or documentation files are not updated.

The commercial license header may contain some additional
lines so that its line count equals with the previous
license header. Reason for this is that some autotests
use hard coded line numbers and a change in the line count
causes failures in tests.

Task-number: QTQAINFRA-4168
Change-Id: I7614c3aa5c94496c9ca7f1baafee14f7b3d1cadd
Reviewed-by: Antti Kokko <antti.kokko@qt.io>