Update commercial license headers
Updated header.COMM to the files in tqtc-qtlottie.
Examples, tests, or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the the line count causes failures in tests.

Task-number: QTQAINFRA-4948
Change-Id: Id44470f0f8cc3d3a3535fd57496019529bff7955
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Kai Koehne <kai.koehne@qt.io>