Move parse() function into a sharable location
This introduces QLottieRoot, which is used to kick off the parsing and
building of the tree. This so that the functionality can be shared
with upcoming tools which won't be using the batch renderer.

Also fixes a bug with image layers, where the image asset was stored
under the wrong object property name and hence never displayed.

Change-Id: I93fc1b959f4d3861e2c725e242457ddeec365a16
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>