Add support for Precomposition layers
Since Precomp layers hold a list of child layers, just like the root
animation object does, the code to parse and construct those child
layer objects has been broken out into a shared function. Precomp
layers can also hold other precomp layers as children, so the parsing
& construction is recursive.

Before this, all layers were direct children of the root object. So
the checking of the start- and end-frame (ip & op) values, to
determine if the layer should be shown or not, could be done by the
renderer (by calling active() on the children of the root).

Since there can now be layers deeper down in the object tree, an
internal isActive property has been added to the layers so that the
state according to the current frame number can be remembered and
heeded in the render() functions.

Also adds a number of bugfixes:
- On missing transform element, default scale and opacity to 1, not 0
- Support transitive linked ("parent") layers
- Ensure correct transform of layer clip path ("matte"), if set
- Support non-0 value of layers' startTime property

Change-Id: I0454cbfba5908ee5604d2956ea3b797cded43f8b
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>