doc: Add a warning about using Qt Lottie with untrusted sources
Qt Lottie Animation is not tested to an extent where we can
confidently say it will handle potentially malicious input.
Experience shows that it does make assumptions about input and
may at least assert if the data is invalid. Until sufficient
testing is made, we document that it should only be used with
trusted data, which was also the original purpose of the module.

Pick-to: 6.8 6.9
Change-Id: I9f2d90c1049e5d9ef31ac48164985870dd166dbb
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>