Support linked ("parent") layers of any index number
The implementation used index number 0 as a special value indicating
no linked layer. But both 0 and negative numbers are possible valid
index numbers in Lottie files. Fix by explicitly storing whether a
layer has a linked layer set or not.

Also changes a variable name from "parentLayer" to "linkedLayer", so
the code consistently uses the latter term for this feature, and
avoids confusion with the parent/child relationships in the element
tree.

Change-Id: I56dd206aebd4d12a851b3f5d42b9040b07eaca1c
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>