Improve detection of animation keyframe schema
The parser would determine to use old- or new-style keyframe schema
based on the version number present in the file. However, this has
proved to be unreliable: it is not uncommon to encounter files in the
wild with misleading or missing version numbers. Such files would
typically fail to render at all.

Instead, we base the detection on the presence of a data value ("s")
in the final keyframe, as that is all that is needed for new-style
parsing.

The effective change is the one-line setting of schemaChanged in
qlottieproperty_p.h. All the rest is just removing the handling of
the file version number info, as that is no longer needed.

Change-Id: Iad1a2441df347e0ec700ecd66bddef5ab83f36d3
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>