lottietoqml: Fix viewbox calculation
Attempting to measure the actual size of the animation does not
work, since it might change from frame to frame.

LottieAnimation uses a view box of (0, 0) to (w, h) where 'w' and
'h' are properties on the root object, so this is the defined viewbox
of the JSON file. Instead of trying to be clever, we just use this.

Pick-to: 6.10
Change-Id: I09674f157ec19c92a051ee3dec49600ef8edd0fa
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>