tests: Set the VectorImage animation time on the scene directly
For Lancelot tests, we currently set the animation frame on
the parser and then get the data at that frame.

First of all, this has exposed some bugs which are only visible
with this approach, not when we are actually running the
animations through VectorImage.

In addition, it does not actually test the animations we
generate, since we are resolving the animations through the
Qt Lottie parser.

To make the results actually reflect what VectorImage would
be rendering, we get all animations that are not part of
a group (so in practice all the top level parallel animations)
and then set the current time of these to match the time
of the frame we want.

Change-Id: Id3759d95423f3ffe1991637056a9e69f46022b27
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>
(cherry picked from commit 83c1743f2cd5c1d934ec765946235c1d95bd848a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>