Improve parsing of easings of animated vector properties
An animated vector property may specify individual easing curves for
each vector element. In that case, the in and out x,y values will
themselves be vectors, with one element for each curve.

The parsing instead interpreted such in/out vectors as consecutive
bezier segments, resulting in wrong rendering.

This patch improves the parsing by instead just taking the easing
curve for the first element. The others are ignored, since individual
element animation is not supported yet, ref. QTBUG-138143

This fixes the rendering of the USAMapWithOutlines.json animation in
the lottietoqml example.

Fixes: QTBUG-138213
Change-Id: Ic230c50231cb4ef2f18f8f90131b5f34f13d41d7
Reviewed-by: Kwanghyo Park <kwanghyo.park@qt.io>
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>