Ensure that LottieAnimation really stops at the requested frame
It can happen that a repaint is already in the queue when the
LottiaAnimation is asked to pause. Since each paint operation would
increase the framenumber, this would change which framenumber was
actually left rendered on screen when the pause started. Fix by
skipping the frame number stepping when pause is enabled.

This helps to ensure deterministic results in the baseline tests.

Change-Id: Ib6255d51b58e3ff7a0552462f23fe54155124e43
Pick-to: 6.10
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>