Make warnings about unsupported features opt-in
It is known and documented that not all features in Lottie are
currently supported. The parser encountering such features in a lottie
file is thus not an unexpected situation that warrants emitting an
unconditional warning.

Instead, this patch downgrades the severity level of these messages to
Info, and sets the default filter level to Warning. In effect, this
silences those messages by default, but allows the user to opt in to
them by enabling (at least) the Info level in the runtime loggin
settings.

This significantly reduces the noise level when opening many lottie
files, and makes it easier to detect the real, unexpected warnings.

Change-Id: Ic6ba75a38b7288a8a19acfc29fb6ee3b63043493
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>
(cherry picked from commit e637a72ca1cdb8f9bf0e72a157d8b51e1761f284)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>