lottietoqml: Generate structure items for groups when needed
When a Shape Group has a transform, and contains other Groups that may
have their own transforms, the resulting combined transform on the
generated Shape object could become wrong, since the two levels of
transforms were combined without possibility to keeping the original
order of transform operations.

Fix by detecting this case, and generate an intermediate QML Item (a
"structure node") container representing the parent Group and having
its transform.

In addition to fixing rendering errors in some cases, this change in
many cases causes significant savings in the size of the generated QML
file, when complex animations on group level need not be repeated for
multiple children, but instead is present only once in the structure
node.

Change-Id: I9ec8f29dddf3e06409d2ac516a2dcf4aef833a3b
Pick-to: 6.10
Fixes: QTBUG-139553
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>