Fix path combinations: Render grouped path elments as a single path
Successive path elements in the same group are intended to be combined
into a single path. This is necessary in order to render paths with
holes (doughnut shapes), using OddEven fill rule.

To implement that, this commit replaces the LottieRenderer interface's
indivdual path element functions with a single function taking a list
of path elements that should be rendered as a combined path.

It also factors out a commen renderChildren() function used in the
render function of all lottie elements that can have children. This
function then builds the appropriate list of path elements to be
passed to the new renderer function.

Fixes: QTBUG-137912
Change-Id: I0345c22b03322ef7cfe40ba4667b08ed87cdb88b
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>