lottietoqml: Don't output full file path in comment
As part of the comment generated by lottietoqml, we would include the
entire file path as it was provided in the arguments. With the
introduction of the automatic svgtoqml build step, this argument
will now include the full absolute path.

Since the QML file is embedded in the binary, this caused internal
file paths from our build farm to leak into the binaries, triggering
an alert in the QA system.

Change-Id: I77bcb53bd475a0418a9fbc5075a6efb54a3ffc14
Reviewed-by: Samuli Piippo <samuli.piippo@qt.io>