Adapt to changes in VectorImage generator
Since the mask support was added to the generator, it now
expects us to provide a unique ID for the node. This isn't
actually used for anything so far, so we just provide one
and keep track of them based on the node pointer.

One special case is the "layer rect" shape which does not
correspond to a specific shape node, so the node pointer
is null there. This will never be referenced by an animation
anyway, so we just leave the ID empty for it.

Task-number: QTBUG-121537
Change-Id: Iffa0bb4d1515a94f23da506c0c83160b036db514
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>