Make vector image backend depend on Helpers module
In order for the generated code in VectorImage to be correctly
resolved on Android, we need to make the dependency on the
Helpers module explicit. In order to do this, we add a dependency
from the generator library (which the backend plugin will already
depend on) to the helpers library. In order for this to work, the
helpers module can't be NO_PLUGIN_OPTIONAL, but I am not sure why
this was set to begin with. The helpers module plugin code is
generated, so there should be no need to depend on it.

Task-number: QTBUG-140187
Change-Id: Ie7229e924b57ce5cf2285454ff9ca18edee74007
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>
(cherry picked from commit 7503bf4f86eb8ecc238940b473bf281c36f443cc)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>