lottietoqml: Simplify layer and linked-layer handling
With the recent changes, we are now generating all ids for the
generated qml code ourselves, in the visitor. That means we no longer
need the special handling of layers and layer "parents" / links
(transform references), so we simplify that away.

This commit is needed to adapt to changes in the generator api.

Change-Id: I0c4c11a1a2c969af128709d571170ef482a4a734
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>