test: Check that VectorImage is able to load Lottie files
This test goes through the test files under baseline tests
and tries to load each one of them using VectorImage,
reporting if any of them fail to parse.

We also introduce a test to verify that a broken file does
not parse correctly and that we only support loading Lottie
when assumeTrustedSource is set to true.

Fixes: QTBUG-140812
Pick-to: 6.10
Change-Id: Ieea4b640b8623ee94b5b5d3704720aa4d45cc590
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>