Fix lottietoqml invocation when there is no windowing system
Amends 184ddaefcba14861e65cc00be35d3ca14c23eaf9. This change
fixed an issue when cross-compiling and a platform plugin
is not available. It selects the offscreen plugin when this
case is detected, but uses the default platform plugin
otherwise, since the offscreen plugin has limited support for
fonts so we do not want it as the default.

However, this did not fix the case where the platform plugin
exists but there is no windowing system, e.g. when building
over SSH.

Luckily, the minimal QPA plugin does have proper font
support and also works over SSH. We can therefore use this
as the default as long as it exists. The only exception is
if we need GUI to preview the result (if -v is specified or
if there is no output file.)

Fixes: QTBUG-140675
Change-Id: I465c660f8bc2f7e1961cecff1de64a1531cdc12f
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>