Baseline test: Fix image grabbing size for eglfs
Running on eglfs, the grabbed image will always be the entire screen
size, independently of the size of the rendered scene. The
scenegrabber code tries to crop away the unused parts.

The recent improvements to Main.qml had an unwanted side effect: it
broke the scene size calculation, causing all grabbed images to be
cropped to only 30x20 pixels. Fix by using a better approach to
calculate the scene size.

Change-Id: I1d3b74442f175982c545b6ceca8b7bbc92482062
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>