Cleanup naming regarding mattes
In the implementation, qtlottie many places used the terms "mask" and
"clipping" when dealing with the feature known as "matte" in
Lottie. This is confusing, since Lottie has a different, similar
feature it calls "masks" (currently not supported in qtlottie). This
commit cleans up the naming and language to avoid the confusion.

Task-number: QTBUG-137913
Change-Id: Id430d00a34676e0958741d5a2171e99e74bec7c9
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>
Reviewed-by: Kwanghyo Park <kwanghyo.park@qt.io>