Add layerSize() function
Not all layer types have an explicit size defined; they are implicitly
the size of their parent layer, or ultimately, the animation root
itself. Add a utility function to query the the size, independent
of it being explicitly or implicitly specified. Useful for
implementing some layer-scope features, e.g. mattes.

Change-Id: Ibe07c858d1127cabd77a4a28fa5309bbc9b4b6be
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>