LottieAnimation: Don't assert when setting source to empty
When setting the source URL to an empty string, we will get
a QQmlFile with Status == Null, but in Lottie we would assert
that it is either Error or Ready.

We catch this before trying to load the file and set our own
status back to Null when this happens.

Pick-to: 6.8 6.5
Change-Id: I8b4c5ecd37ecaed3e79c7077d76c9408989abf46
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>
(cherry picked from commit 921f83bc472fd919fcdcd1b9014223028c2ef1e3)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>