VectorImage: Implement motion along path
Positions can have curved paths in Lottie. This is
implemented by setting tangents on the end position for the
frame, and Lottie builds up a bezier path for the movement
internally.

Since we build this path even if all the content is linear, we
first check if there are any curves present. If there are,
we apply a special case where we set the motion path for the
movement instead of a transform animation.

Task-number: QTBUG-140751
Change-Id: I4410572428ff9124e7b103f74e3977b7dd6572b2
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>