lottietoqml: unify timepoint calculations and fix time offsets
Factor out a common function to convert from Lottie frame number to
qml generator time points. This also ensures that layer time offsets
are applied consistently, which was not the case.

Fix time offset handling in general: Only applies to precomp layers,
and time points should be wrapped if the resulting time point falls
outside the animation duration.

Correct the layer start and end point handling: Should not be subject
to the layer's own time offset. These time points should not be
wrapped, so need to handle exceeding numbers manually.

Pick-to: 6.11
Fixes: QTBUG-142319
Change-Id: Icefd6ee0759e257e78941581be38f457c5fc025c
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>