lottietoqml: simplify and correct keyframe transcoding
Utilize that we can now rely on getting one Curve object for every
keyframe, including the final special one, with valid startFrame and
startValue. This facilitates simplifying the recoding of the keyframe
data to the qml generator's format. Must only handle that the Curve
has the easing for the coming animation step, while the generator
keyframe expoects it for the preceding step.

Drop storing a special start value for frame 0 in the collect()
functions: it can cause wrong results in case of layer time offset,
and it is anyway done correctly by the fill() functions.

Task-number: QTBUG-142319
Change-Id: I238d7f15440cfff3b73e0ce5374fcdac2fe16861
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>