Unify keyframe interpretation across lottie versions
When reading old schema lottie files, animation keyframes are stored
as one curve object for each, including the final, special one that
just marks the end frame. That one is stored with startFrame==endFrame
and startValue==endValue, and no easing, and the complete flag false.

New schema files were read differently, and the special last curve was
missing. This difference caused needless complexity when
interpretating the stored curves in lottietoqml.

This commit fixes the issue with a minimal trick to change the new
schema reading to also produce the special last curve. Thus the stored
curves will be the same, independently of lottie schema.

Task-number: QTBUG-142319
Change-Id: I665171c0e658f4f3345acd84bb72fb3563be162e
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>