Fix determination of static or animated property during parsing
Although the spec says that properties should have an "a" value to
flag if the property is animated or not, this is many times missing in
real world animations. As Qt releied on the presence of the flag, it
would sometimes fail to parse animated properties as intended. Fix by
using a simple heuristic to determine if the property has a static or
animated value.

As a driveby, clean up a bit the conversion from json, avoiding
redundant conversions.

Change-Id: Id32ec2537b9170e879ae7ffef382cb5330af9d67
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>
(cherry picked from commit 82f0ead92398a91be98c2793e8dd18a5ca5015ae)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>